# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      description: Validate admin password and return success
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredential"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Invalid password
...
  /admin/analytics:
    get:
      operationId: getAdminAnalytics
      tags: [admin]
      summary: Get analytics overview
      description: Returns totals, growth stats, DAU/WAU/MAU, and activity data
      security:
        - AdminAuth: []
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAnalytics"
        "401":
          description: Unauthorized
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: List all users
      description: Returns all registered users with activity counts and block status
      security:
        - AdminAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
  /admin/users/{email}/detail:
    get:
      operationId: getAdminUserDetail
      tags: [admin]
      summary: Get full data for a single user
      description: Returns profile, groups, expenses, settlements, IET, and business trips
      security:
        - AdminAuth: []
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
        "401":
          description: Unauthorized
        "404":
          description: User not found
  /admin/users/{email}/block:
    post:
      operationId: blockAdminUser
      tags: [admin]
      summary: Block a user
      security:
        - AdminAuth: []
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBlockUserBody"
      responses:
        "200":
          description: User blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminActionResult"
        "401":
          description: Unauthorized
        "404":
          description: User not found
  /admin/users/{email}/unblock:
    post:
      operationId: unblockAdminUser
      tags: [admin]
      summary: Unblock a user
      security:
        - AdminAuth: []
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User unblocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminActionResult"
        "401":
          description: Unauthorized
        "404":
          description: User not found
  /admin/users/{email}:
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Delete a user account
      security:
        - AdminAuth: []
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminActionResult"
        "401":
          description: Unauthorized
        "404":
          description: User not found
  /admin/users/{email}/export:
    get:
      operationId: exportAdminUser
      tags: [admin]
      summary: Export all data for a user
      security:
        - AdminAuth: []
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User data export
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized
        "404":
          description: User not found
  /admin/performance:
    get:
      operationId: getAdminPerformance
      tags: [admin]
      summary: Get server performance metrics
      security:
        - AdminAuth: []
      responses:
        "200":
          description: Performance data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPerformance"
        "401":
          description: Unauthorized
  /admin/audit-logs:
    get:
      operationId: getAdminAuditLogs
      tags: [admin]
      summary: Get audit log entries
      security:
        - AdminAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: action
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogPage"
        "401":
          description: Unauthorized
  /admin/trips:
    get:
      operationId: getAdminTrips
      tags: [admin]
      summary: List all legacy trips
      security:
        - AdminAuth: []
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminTrip"
        "401":
          description: Unauthorized
components:
  securitySchemes:
    AdminAuth:
      type: http
      scheme: bearer
  schemas:
...
    AdminCredential:
      type: object
      required: [password]
      properties:
        password:
          type: string
    AdminLoginResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    AdminBlockUserBody:
      type: object
      properties:
        reason:
          type: string
    AdminActionResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    AdminAnalytics:
      type: object
      required:
        - totalUsers
        - newUsersToday
        - newUsersThisWeek
        - blockedUsers
        - verifiedUsers
        - dau
        - wau
        - mau
        - totalGroups
        - totalIetExpenses
        - totalBusinessTrips
        - signupsByDay
        - topGroupOwners
        - expensesToday
        - settledAmountToday
        - ietToday
        - serverErrors
        - uptimeSeconds
        - dauUsers
        - wauUsers
      properties:
        totalUsers:
          type: integer
        newUsersToday:
          type: integer
        newUsersThisWeek:
          type: integer
        blockedUsers:
          type: integer
        verifiedUsers:
          type: integer
        dau:
          type: integer
        wau:
          type: integer
        mau:
          type: integer
        totalGroups:
          type: integer
        totalIetExpenses:
          type: integer
        totalBusinessTrips:
          type: integer
        signupsByDay:
          type: array
          items:
            $ref: "#/components/schemas/DayCount"
        topGroupOwners:
          type: array
          items:
            $ref: "#/components/schemas/GroupOwnerStat"
        expensesToday:
          type: integer
        settledAmountToday:
          type: number
        ietToday:
          type: integer
        serverErrors:
          type: integer
        uptimeSeconds:
          type: integer
        dauUsers:
          type: array
          items:
            $ref: "#/components/schemas/ActiveUser"
        wauUsers:
          type: array
          items:
            $ref: "#/components/schemas/ActiveUser"
    DayCount:
      type: object
      required: [date, count]
      properties:
        date:
          type: string
        count:
          type: integer
    GroupOwnerStat:
      type: object
      required: [email, groupCount]
      properties:
        email:
          type: string
        groupCount:
          type: integer
    ActiveUser:
      type: object
      required: [email, lastSeen]
      properties:
        email:
          type: string
        lastSeen:
          type: string
    AdminUser:
      type: object
      required: [id, email, createdAt, isBlocked, groupsOwned, groupMemberships, ietExpenses, businessTrips]
      properties:
        id:
          type: integer
        email:
          type: string
        createdAt:
          type: string
        isBlocked:
          type: boolean
        lastLoginAt:
          type: string
          nullable: true
        groupsOwned:
          type: integer
        groupMemberships:
          type: integer
        ietExpenses:
          type: integer
        businessTrips:
          type: integer
    AdminUserDetail:
      type: object
      required: [isBlocked, profile, groups, groupMemberships, expenses, settlements, ietExpenses, businessTrips]
      properties:
        isBlocked:
          type: boolean
        lastLoginAt:
          type: string
          nullable: true
        profile:
          $ref: "#/components/schemas/AdminUserProfile"
        groups:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserGroup"
        groupMemberships:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserGroupMembership"
        expenses:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserExpense"
        settlements:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserSettlement"
        ietExpenses:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserIetExpense"
        businessTrips:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserBusinessTrip"
    AdminUserProfile:
      type: object
      required: [name, upiId, travelStyle, avatar]
      properties:
        name:
          type: string
        upiId:
          type: string
        travelStyle:
          type: string
        avatar:
          type: string
    AdminUserGroup:
      type: object
      required: [id, name, emoji, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        emoji:
          type: string
        createdAt:
          type: string
    AdminUserGroupMembership:
      type: object
      required: [groupId, groupName, memberName]
      properties:
        groupId:
          type: string
        groupName:
          type: string
        memberName:
          type: string
    AdminUserExpense:
      type: object
      required: [id, title, amount, category, date, groupName]
      properties:
        id:
          type: string
        title:
          type: string
        amount:
          type: number
        category:
          type: string
        date:
          type: string
        groupName:
          type: string
    AdminUserSettlement:
      type: object
      required: [id, amount, date, mode, groupName]
      properties:
        id:
          type: string
        amount:
          type: number
        date:
          type: string
        mode:
          type: string
        groupName:
          type: string
...
    AdminPerformance:
      type: object
      required: [uptimeSeconds, memory, requests, slowRequests, topRoutes]
      properties:
        uptimeSeconds:
          type: integer
        memory:
          type: object
          required: [heapUsedMb, heapTotalMb, rssMb]
          properties:
            heapUsedMb:
              type: integer
            heapTotalMb:
              type: integer
            rssMb:
              type: integer
        requests:
          type: object
          required: [total, errors, avgResponseMs, p95ResponseMs]
          properties:
            total:
              type: integer
            errors:
              type: integer
            avgResponseMs:
              type: integer
            p95ResponseMs:
              type: integer
        slowRequests:
          type: array
          items:
            $ref: "#/components/schemas/SlowRequest"
        topRoutes:
          type: array
          items:
            $ref: "#/components/schemas/RouteStats"
...
    AuditLogPage:
      type: object
      required: [total, page, limit, logs]
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
...
    AdminTrip:
      type: object
      required: [id, tagNumber, name, emoji, ownerEmail, memberCount, createdAt]
      properties:
        id:
          type: integer
        tagNumber:
          type: string
        name:
          type: string
        emoji:
          type: string
        ownerEmail:
          type: string
        memberCount:
          type: integer
        createdAt:
          type: string